requirements.txt
certifi==2021.10.8
ciso8601==2.2.0
click==8.0.4
Flask==2.0.3
influxdb-client==1.1.0
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.0
minimalmodbus==1.0.2
numpy==1.22.3
paho-mqtt==1.6.1
pandas==1.4.1
pyserial==3.4
python-dateutil==2.8.2
pytz==2021.3
Rx==3.2.0
six==1.16.0
urllib3==1.26.9
Werkzeug==2.0.3
